server-side/requirements.txt

amqp==5.2.0
ansicon==1.89.0
anyio==4.3.0
arrow==1.3.0
asgiref==3.7.2
asyncio==3.4.3
billiard==4.2.0
openai
black==24.4.2
blessed==1.20.0
blinker==1.7.0
boto3==1.28.68
botocore==1.31.68
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
decorator==4.4.2
defusedxml==0.7.1
dj-stripe==2.8.3
Django==4.2.6
django-allauth==0.61.1
django-celery-results==2.5.1
django-compat==1.0.15
django-cors-headers==4.3.0
django-picklefield==3.2
django-q==1.3.9
django-storages==1.14.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
ffmpeg==1.4
filelock==3.13.3
firebase-admin==6.5.0
flake8==7.0.0
Flask==3.0.2
Flask-Cors==4.0.0
flower==2.0.1
fsspec==2024.3.1
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gradio_client==0.15.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
humanize==4.9.0
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
jinxed==1.2.1
jmespath==1.0.1
kombu==5.3.6
MarkupSafe==2.1.5
mccabe==0.7.0
moviepy==1.0.3
msgpack==1.0.8
mypy-extensions==1.0.0
mysqlclient==2.2.4
ngrok==1.1.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.3.0
platformdirs==4.2.2
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pyparsing==3.1.2
pypdfium2==4.30.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
stripe==4.2.0
tornado==6.4
tqdm==4.66.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.8
websockets==11.0.3
Werkzeug==3.0.2
youtube-dl==2021.12.17
google-api-python-client
google-auth 
google-auth-oauthlib 
google-auth-httplib2